# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "天气管理系统后端API文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:9027",
    "basePath": "/",
    "paths": {
...
        "/node": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "获取所有设备信息",
                "responses": {
                    "200": {
                        "description": "返回设备列表",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "管理员新建设备",
                "parameters": [
                    {
                        "description": "新设备",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Node"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    },
                    "442": {
                        "description": "设备已存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        },
        "/node/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "展示当前用户可见设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前用户的ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回设备列表",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "管理员修改设备信息",
                "parameters": [
                    {
                        "description": "新设备",
                        "name": "node",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Node"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "具体修改的设备的ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "设备不存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "管理员删除设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "具体删除的设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    },
                    "400": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        },
        "/node/setTime": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "管理员修改设备系统时间",
                "parameters": [
                    {
                        "type": "string",
                        "description": "yy.mm.dd hh:mm:ss",
                        "name": "time",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    },
                    "442": {
                        "description": "Server无响应",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        },
...
        "/phy/:id": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "物理量管理"
                ],
                "summary": "管理员修改物理量信息",
                "parameters": [
                    {
                        "description": "物理量信息",
                        "name": "node",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Physical"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "具体修改的物理量的ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "物理量不存在",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.SwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "msg": {
                    "description": "描述",
                    "type": "string"
                }
            }
        },
        "controller.SwaggerResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "msg": {
                    "description": "描述",
                    "type": "string"
                }
            }
        },
        "model.Node": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nodeName": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                }
            }
        },
        "model.Physical": {
            "type": "object",
            "properties": {
                "chinese_name": {
                    "type": "string"
                },
                "conversion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meaning": {
                    "type": "string"
                },
                "physicalName": {
                    "type": "string"
                }
            }
        },
...
    }
}